# lib/api-zod/src/generated/types/containerSystemAudit.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */

export interface ContainerSystemAudit {
  id: number;
  /** @nullable */
  recordId?: number | null;
  kind: string;
  action: string;
  /** @nullable */
  beforePayload?: string | null;
  /** @nullable */
  afterPayload?: string | null;
  /** @nullable */
  actorId?: number | null;
  createdAt: string;
}