# lib/api-zod/src/generated/types/containerSystemRecord.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerSystemRecordPayload } from './containerSystemRecordPayload';

export interface ContainerSystemRecord {
  id: number;
  kind: string;
  status: string;
  reference: string;
  payload: ContainerSystemRecordPayload;
  /** @nullable */
  createdBy?: number | null;
  createdAt: string;
  updatedAt: string;
}