# lib/api-zod/src/generated/types/containerSystemRecordInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerSystemRecordInputPayload } from './containerSystemRecordInputPayload';

export interface ContainerSystemRecordInput {
  kind: string;
  status?: string;
  payload: ContainerSystemRecordInputPayload;
}