# lib/api-zod/src/generated/types/containerSystemRecordUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerSystemRecordUpdatePayload } from './containerSystemRecordUpdatePayload';

export interface ContainerSystemRecordUpdate {
  status?: string;
  payload?: ContainerSystemRecordUpdatePayload;
}