# lib/api-zod/src/generated/types/containerSystemSnapshot.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Sabaik Almasa API
 * OpenAPI spec version: 0.1.0
 */
import type { ContainerSystemRecord } from './containerSystemRecord';
import type { ContainerSystemSummary } from './containerSystemSummary';

export interface ContainerSystemSnapshot {
  summary: ContainerSystemSummary;
  records: ContainerSystemRecord[];
  expiringContracts: ContainerSystemRecord[];
  recent: ContainerSystemRecord[];
}